Remove users table component
